feat(console): console-family groundwork for SD/Quantum support
Prepares the app to drive DiGiCo SD and Quantum consoles alongside the
S21/S31 it already speaks to. Nothing changes for S-series users except
one deliberate fix, noted below; the SD/Quantum connection path itself is
the next phase.

Shared plumbing, previously duplicated or S-series-only:

* ConsoleTx (console/console_tx.rs) becomes the single console write path,
  replacing the "GP OSC first, iPad-protocol fallback" block that had been
  copy-pasted across eleven sites in five engines. send_util.rs is absorbed
  into it and removed.

* SentLog, alongside it, records what we send. The iPad link echoes our own
  engine writes back at us, and an absolute-mode gang would happily copy a
  snapshot recall's echo across every member; the inbound chain now screens
  those echoes out of gang and pan-link propagation and out of the
  automation-override registry.

* console/inbound.rs holds the inbound side-effect chain — mirror update,
  dirty screening, operator override, gang, pan link, macro learn — which
  until now ran only in the GP OSC loop. It serves the GP loop, the Mode 2
  iPad mirror and the Mode 3 proxy capture alike. This is the deliberate
  fix: desk moves seen over the iPad protocol drive gangs and pan link,
  which they previously did not.

* Channel numbers widen from u8 to u16 (ChannelId, ConsoleConfig counts,
  send and bus indices, pan-link bindings, the monitor and web stacks). A
  Quantum7 reports 256 input channels. EQ, dynamics and GEQ band indices
  stay u8, as do GP OSC wire numbers: those are processing-structure
  indices, not console-size dependent.

* osc::decode_udp_tolerant pads and retries datagrams whose length is not a
  multiple of four. SD and Quantum emit them; strict decoders reject them.

The family model itself:

* model/family.rs adds ConsoleFamily, PadQuirks, ConsoleProfile and
  ParamSupport. Capacities are deliberately absent — channel counts and bus
  modes stay discovered at runtime from the console, so one profile covers a
  whole range whose models differ in size and licensed channel count. The
  S-series quirk values are hardware-verified; the SD/Quantum ones are
  flagged in code as hypotheses pending a hardware probe, and every one is
  overridable at runtime rather than requiring a rebuild.

* The Pad codec is now quirk-driven: to_pad_suffix(&PadQuirks),
  from_pad_suffix, to_pad_path_prefix and encode_pad_parameter are the real
  implementations. The old *_ipad_* names remain as inline wrappers pinned
  to the S21 quirks, so the existing golden tests keep pinning S21 wire
  strings without being edited. Production code never calls a wrapper.

* ParameterPath::support(family) classifies each parameter per family and
  gates applicable_to_for_family and available_for_channel_on. Everything is
  Verified on S-series, so S21 behaviour is unchanged by construction.

* ConsoleConfig gains family and pad_quirk_overrides, taking the show file
  to v19; older files load untouched. AppPreferences gains matching seeds
  for new shows. The recall-scope editor hides on families whose consoles
  have no such screen.

Two robustness fixes fall out of the same work. decode_udp_tolerant now
requires a type-tag marker before retrying a padded decode: without that
guard it could swallow a DiGiCo bare-path packet carrying a single-character
inline value, reading the value byte as a type tag and silently dropping the
update that parse_digico_packet would have recovered. And to_gp_osc_number
does its bus-offset arithmetic with checked_add, since the Macros tab parses
an unbounded channel number from free text — previously that could panic in
debug and, worse, wrap into a plausible-looking channel number in release.

795 tests pass and clippy is clean. New coverage includes a 48-combination
encode/parse round-trip proving the codec stays self-inverse under any
family's quirks, an invariant that nothing marked usable lacks a Pad path,
and show-file load tests for both a legacy v18 and a Quantum v19.

+++ src/console/discovery.rs
@@ -11,12 +11,12 @@ use crate::model::config::{ChannelMode, ConsoleConfig, PlusMode};
 /// exactly one master and the config has no field for it.
 pub fn apply_channel_counts(
     config: &mut ConsoleConfig,
-    inputs: u8,
-    aux: u8,
-    groups: u8,
-    control_groups: u8,
-    matrices: u8,
-    _master: u8,
+    inputs: u16,
+    aux: u16,
+    groups: u16,
+    control_groups: u16,
+    matrices: u16,
+    _master: u16,
 ) {
     config.input_channel_count = inputs;
     config.aux_output_count = aux;
@@ -36,7 +36,7 @@ pub fn apply_channel_counts(
 
 /// Map a channel type name (from /console/channel/counts/{type}) to a config update.
 /// Returns true if the type was recognized and applied.
-pub fn apply_channel_count(config: &mut ConsoleConfig, channel_type: &str, count: u8) -> bool {
+pub fn apply_channel_count(config: &mut ConsoleConfig, channel_type: &str, count: u16) -> bool {
     match channel_type {
         "input" => {
             config.input_channel_count = count;
